First attempt at a polyroute directive
It uses the same logic as a polyline but requests a route between the the polyline points. Its uses the direction service and its waypoints so there is one call for every 10 point on the polyline. More than 10 points breaks into seperate routes. Each returned route is turned into a polyline with all the point in the route, these polyline are not directly editable by the user. The path of the polyroute is in essence all the waypoints for the final route.